refactor: modularize endpoint-page (#3594)
