+++ src/app/(dashboard)/dashboard/endpoint/EndpointPageClient.tsx
@@ -13,136 +13,26 @@ import McpDashboardPage from "./components/MCPDashboard";
 import NotionSourceCard from "./components/NotionSourceCard";
 import VscodeTokenAliasCard from "./VscodeTokenAliasCard";
 
-const BUILD_TIME_CLOUD_URL = process.env.NEXT_PUBLIC_CLOUD_URL || null;
-const CLOUD_ACTION_TIMEOUT_MS = 15000;
-
-type TranslationValues = Record<string, string | number | boolean | Date>;
-type CloudflaredTunnelPhase =
-  | "unsupported"
-  | "not_installed"
-  | "stopped"
-  | "starting"
-  | "running"
-  | "error";
-
-type CloudflaredTunnelStatus = {
-  supported: boolean;
-  installed: boolean;
-  managedInstall: boolean;
-  installSource: string | null;
-  binaryPath: string | null;
-  running: boolean;
-  pid: number | null;
-  publicUrl: string | null;
-  apiUrl: string | null;
-  targetUrl: string;
-  phase: CloudflaredTunnelPhase;
-  lastError: string | null;
-  logPath: string;
-};
-
-type TailscaleTunnelPhase =
-  | "unsupported"
-  | "not_installed"
-  | "needs_login"
-  | "stopped"
-  | "running"
-  | "error";
-
-type TailscaleTunnelStatus = {
-  supported: boolean;
-  installed: boolean;
-  managedInstall: boolean;
-  installSource: string | null;
-  binaryPath: string | null;
-  loggedIn: boolean;
-  daemonRunning: boolean;
-  running: boolean;
-  enabled: boolean;
-  tunnelUrl: string | null;
-  apiUrl: string | null;
-  phase: TailscaleTunnelPhase;
-  platform: string;
-  brewAvailable: boolean;
-  lastError: string | null;
-  pid: number | null;
-};
-
-type NgrokTunnelPhase =
-  | "unsupported"
-  | "not_installed"
-  | "stopped"
-  | "needs_auth"
-  | "starting"
-  | "running"
-  | "error";
-
-type NgrokTunnelStatus = {
-  supported: boolean;
-  installed: boolean;
-  running: boolean;
-  publicUrl: string | null;
-  apiUrl: string | null;
-  targetUrl: string;
-  phase: NgrokTunnelPhase;
-  lastError: string | null;
-};
-
-type TunnelNotice = {
-  type: "success" | "error" | "info";
-  message: string;
-};
-
-type APIPageClientProps = {
-  machineId: string;
-};
-
-type EndpointProviderSummary = {
-  id: string;
-  provider: {
-    name: string;
-    alias?: string;
-  };
-};
-
-type EndpointModelSummary = {
-  id: string;
-  owned_by?: string;
-  parent?: string;
-  type?: string;
-  custom?: boolean;
-  root?: string;
-};
-
-type CopyHandler = (text: string, key?: string) => void | Promise<void>;
-
-type EndpointTunnelVisibility = {
-  showCloudflaredTunnel: boolean;
-  showTailscaleFunnel: boolean;
-  showNgrokTunnel: boolean;
-};
-
-type EndpointTab = "apis" | "mcp" | "a2a" | "context-sources";
-
-const ENDPOINT_TABS: Array<{ value: EndpointTab; label: string; icon: string }> = [
-  { value: "apis", label: "APIs", icon: "api" },
-  { value: "mcp", label: "MCP", icon: "extension" },
-  { value: "a2a", label: "A2A", icon: "hub" },
-  { value: "context-sources", label: "Context Sources", icon: "database" },
-];
-
-const DEFAULT_TUNNEL_VISIBILITY: EndpointTunnelVisibility = {
-  showCloudflaredTunnel: true,
-  showTailscaleFunnel: true,
-  showNgrokTunnel: true,
-};
-
-function runEndpointBackgroundTask(taskName: string, task: () => Promise<unknown>) {
-  void task().catch((error) => {
-    console.log("Error running endpoint background task:", taskName, error);
-  });
-}
-
+import { EndpointCard } from "./EndpointCard";
+import { ProviderModelsModal } from "./ProviderModelsModal";
+import {
+  BUILD_TIME_CLOUD_URL,
+  CLOUD_ACTION_TIMEOUT_MS,
+  ENDPOINT_TABS,
+  DEFAULT_TUNNEL_VISIBILITY,
+  runEndpointBackgroundTask,
+  type TranslationValues,
+  type CloudflaredTunnelStatus,
+  type TailscaleTunnelStatus,
+  type NgrokTunnelStatus,
+  type TunnelNotice,
+  type APIPageClientProps,
+  type EndpointProviderSummary,
+  type EndpointModelSummary,
+  type CopyHandler,
+  type EndpointTunnelVisibility,
+  type EndpointTab,
+} from "./helpers";
 export default function APIPageClient({ machineId }: Readonly<APIPageClientProps>) {
   const [resolvedMachineId, setResolvedMachineId] = useState(machineId || "");
   const t = useTranslations("endpoint");
@@ -2259,304 +2149,6 @@ export default function APIPageClient({ machineId }: Readonly<APIPageClientProps
           copied={copied}
           onClose={() => setSelectedProvider(null)}
         />
-      )}
-    </div>
-  );
-}
-
-// -- Sub-component: Provider Models Modal ------------------------------------------
-
-function ProviderModelsModal({
-  provider,
-  models,
-  copy,
-  copied,
-  onClose,
-}: Readonly<{
-  provider: EndpointProviderSummary;
-  models: EndpointModelSummary[];
-  copy: CopyHandler;
-  copied?: string | null;
-  onClose: () => void;
-}>) {
-  const t = useTranslations("endpoint");
-  const tc = useTranslations("common");
-  // Get provider alias for matching models
-  // Filter out parent models (models with parent field set) to avoid showing duplicates
-  const providerAlias = provider.provider.alias || provider.id;
-  const providerModels = useMemo(() => {
-    return models.filter(
-      (m) => !m.parent && (m.owned_by === providerAlias || m.owned_by === provider.id)
-    );
-  }, [models, providerAlias, provider.id]);
-
-  const chatModels = providerModels.filter((m) => !m.type);
-  const embeddingModels = providerModels.filter((m) => m.type === "embedding");
-  const imageModels = providerModels.filter((m) => m.type === "image");
-
-  const renderModelGroup = (title, icon, groupModels) => {
-    if (groupModels.length === 0) return null;
-    return (
-      <div className="mb-4">
-        <h4 className="text-xs font-semibold text-text-muted uppercase tracking-wider mb-2 flex items-center gap-1.5">
-          <span className="material-symbols-outlined text-sm">{icon}</span>
-          {title} ({groupModels.length})
-        </h4>
-        <div className="flex flex-col gap-1">
-          {groupModels.map((m) => {
-            const copyKey = `modal-${m.id}`;
-            return (
-              <div
-                key={m.id}
-                className="flex items-center gap-2 px-3 py-2 rounded-lg hover:bg-surface/60 group"
-              >
-                <code className="text-sm font-mono flex-1 truncate">{m.id}</code>
-                {m.custom && (
-                  <span className="text-[10px] px-1.5 py-0.5 rounded bg-primary/10 text-primary">
-                    {t("custom")}
-                  </span>
-                )}
-                <button
-                  onClick={() => copy(m.id, copyKey)}
-                  className="p-1 hover:bg-sidebar rounded text-text-muted hover:text-primary opacity-0 group-hover:opacity-100 transition-opacity"
-                  title={tc("copy")}
-                >
-                  <span className="material-symbols-outlined text-sm">
-                    {copied === copyKey ? "check" : "content_copy"}
-                  </span>
-                </button>
-              </div>
-            );
-          })}
-        </div>
-      </div>
-    );
-  };
-
-  return (
-    <Modal
-      isOpen
-      onClose={onClose}
-      title={t("providerModelsTitle", { provider: provider.provider.name })}
-    >
-      <div className="max-h-[60vh] overflow-y-auto">
-        {providerModels.length === 0 ? (
-          <p className="text-sm text-text-muted py-4 text-center">{t("noModelsForProvider")}</p>
-        ) : (
-          <>
-            {renderModelGroup(t("chat"), "chat", chatModels)}
-            {renderModelGroup(t("embedding"), "data_array", embeddingModels)}
-            {renderModelGroup(t("image"), "image", imageModels)}
-          </>
-        )}
-      </div>
-    </Modal>
-  );
-}
-
-// -- Sub-component: Endpoint Section ------------------------------------------
-
-function EndpointCard({
-  icon,
-  iconColor,
-  iconBg,
-  title,
-  path,
-  models,
-  copy,
-  copied,
-  baseUrl,
-  badge,
-  modelsLoading = false,
-}: Readonly<{
-  icon: string;
-  iconColor: string;
-  iconBg: string;
-  title: string;
-  path: string;
-  models: EndpointModelSummary[] | null;
-  copy: CopyHandler;
-  copied?: string | null;
-  baseUrl: string;
-  badge?: string;
-  modelsLoading?: boolean;
-}>) {
-  const t = useTranslations("endpoint");
-  const copyId = `endpoint_${path}`;
-  const fullUrl = `${baseUrl.replace(/\/v1$/, "")}${path}`;
-
-  return (
-    <div className="border border-border rounded-lg p-3 hover:bg-surface/30 transition-colors flex flex-col gap-2">
-      <div className="flex items-start gap-2.5">
-        <div className={`flex items-center justify-center size-8 rounded-lg ${iconBg} shrink-0`}>
-          <span className={`material-symbols-outlined text-base ${iconColor}`}>{icon}</span>
-        </div>
-        <div className="flex-1 min-w-0">
-          <div className="flex items-center gap-1.5 flex-wrap">
-            <span className="font-semibold text-xs leading-tight">{title}</span>
-            {badge && (
-              <span className="text-[9px] px-1.5 py-0.5 rounded-full border border-border/60 text-text-muted font-medium uppercase tracking-wider leading-none">
-                {badge}
-              </span>
-            )}
-          </div>
-          <span className="text-xs text-text-muted mt-0.5 block">
-            {models === null
-              ? "—"
-              : modelsLoading
-                ? "..."
-                : t("modelsCount", { count: models.length })}
-          </span>
-        </div>
-      </div>
-      <div className="flex items-center gap-1.5">
-        <code className="flex-1 text-[10px] font-mono text-text-muted bg-surface/80 px-2 py-1 rounded truncate">
-          {path}
-        </code>
-        <button
-          onClick={() => void copy(fullUrl, copyId)}
-          className="shrink-0 flex items-center justify-center size-6 rounded hover:bg-sidebar transition-colors"
-          title={t("copyUrl")}
-        >
-          <span className="material-symbols-outlined text-[12px] text-text-muted">
-            {copied === copyId ? "check" : "content_copy"}
-          </span>
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function EndpointSection({
-  icon,
-  iconColor,
-  iconBg,
-  title,
-  path,
-  description,
-  models,
-  expanded,
-  onToggle,
-  copy,
-  copied,
-  baseUrl,
-  modelsLoading = false,
-}: Readonly<{
-  icon: string;
-  iconColor: string;
-  iconBg: string;
-  title: string;
-  path: string;
-  description: string;
-  models: EndpointModelSummary[];
-  expanded: boolean;
-  onToggle: () => void;
-  copy: CopyHandler;
-  copied?: string | null;
-  baseUrl: string;
-  modelsLoading?: boolean;
-}>) {
-  const t = useTranslations("endpoint");
-  const grouped = useMemo(() => {
-    const map = {};
-    for (const m of models) {
-      const owner = m.owned_by || "unknown";
-      if (!map[owner]) map[owner] = [];
-      map[owner].push(m);
-    }
-    return Object.entries(map).sort((a: any, b: any) => b[1].length - a[1].length);
-  }, [models]);
-
-  const resolveProvider = (id) => AI_PROVIDERS[id] || getProviderByAlias(id);
-  const providerColor = (id) => resolveProvider(id)?.color || "#888";
-  const providerName = (id) => getProviderDisplayName(id, resolveProvider(id));
-  const copyId = `endpoint_${path}`;
-
-  return (
-    <div className="border border-border rounded-lg overflow-hidden">
-      {/* Header (always visible) */}
-      <button
-        onClick={onToggle}
-        className="w-full flex items-center gap-3 p-4 hover:bg-surface/50 transition-colors text-left"
-      >
-        <div className={`flex items-center justify-center size-10 rounded-lg ${iconBg} shrink-0`}>
-          <span className={`material-symbols-outlined text-xl ${iconColor}`}>{icon}</span>
-        </div>
-        <div className="flex-1 min-w-0">
-          <div className="flex items-center gap-2">
-            <span className="font-semibold text-sm">{title}</span>
-            <span className="text-xs px-2 py-0.5 rounded-full bg-surface text-text-muted font-medium">
-              {modelsLoading ? "..." : t("modelsCount", { count: models.length })}
-            </span>
-          </div>
-          <p className="text-xs text-text-muted mt-0.5">{description}</p>
-        </div>
-        <span
-          className={`material-symbols-outlined text-text-muted text-lg transition-transform ${expanded ? "rotate-180" : ""}`}
-        >
-          expand_more
-        </span>
-      </button>
-
-      {/* Expanded content */}
-      {expanded && (
-        <div className="border-t border-border px-4 pb-4">
-          {/* Endpoint path + copy */}
-          <div className="flex items-center gap-2 mt-3 mb-3">
-            <code className="flex-1 text-xs font-mono text-text-muted bg-surface/80 px-3 py-1.5 rounded-lg truncate">
-              {baseUrl.replace(/\/v1$/, "")}
-              {path}
-            </code>
-            <button
-              onClick={() => copy(`${baseUrl.replace(/\/v1$/, "")}${path}`, copyId)}
-              className="p-1.5 hover:bg-surface rounded-lg text-text-muted hover:text-primary transition-colors shrink-0"
-            >
-              <span className="material-symbols-outlined text-[16px]">
-                {copied === copyId ? "check" : "content_copy"}
-              </span>
-            </button>
-          </div>
-
-          {/* Models grouped by provider */}
-          {modelsLoading ? (
-            <div className="flex items-center gap-2 rounded-lg border border-border/70 bg-surface/40 px-3 py-2 text-xs text-text-muted">
-              <span className="material-symbols-outlined animate-spin text-sm">
-                progress_activity
-              </span>
-              <span>{t("loadingModels")}</span>
-            </div>
-          ) : (
-            <div className="flex flex-col gap-2">
-              {grouped.map(([providerId, providerModels]) => (
-                <div key={providerId}>
-                  <div className="flex items-center gap-2 mb-1">
-                    <div
-                      className="size-2.5 rounded-full shrink-0"
-                      style={{ backgroundColor: providerColor(providerId) }}
-                    />
-                    <span className="text-xs font-semibold text-text-main">
-                      {providerName(providerId)}
-                    </span>
-                    <span className="text-xs text-text-muted">
-                      ({(providerModels as any).length})
-                    </span>
-                  </div>
-                  <div className="ml-5 flex flex-wrap gap-1.5">
-                    {(providerModels as any).map((m) => (
-                      <span
-                        key={m.id}
-                        className="text-xs px-2 py-0.5 rounded-md bg-surface/80 text-text-muted font-mono"
-                        title={m.id}
-                      >
-                        {m.root || m.id.split("/").pop()}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
       )}
     </div>
   );
