+++ src/app/(dashboard)/dashboard/endpoint/ProviderModelsModal.tsx
@@ -1,0 +1,94 @@
+"use client";
+
+import { useMemo } from "react";
+import { Modal } from "@/shared/components";
+import { useTranslations } from "next-intl";
+import type { EndpointProviderSummary, EndpointModelSummary, CopyHandler } from "./helpers";
+
+export function ProviderModelsModal({
+  provider,
+  models,
+  copy,
+  copied,
+  onClose,
+}: Readonly<{
+  provider: EndpointProviderSummary;
+  models: EndpointModelSummary[];
+  copy: CopyHandler;
+  copied?: string | null;
+  onClose: () => void;
+}>) {
+  const t = useTranslations("endpoint");
+  const tc = useTranslations("common");
+  // Get provider alias for matching models
+  // Filter out parent models (models with parent field set) to avoid showing duplicates
+  const providerAlias = provider.provider.alias || provider.id;
+  const providerModels = useMemo(() => {
+    return models.filter(
+      (m) => !m.parent && (m.owned_by === providerAlias || m.owned_by === provider.id)
+    );
+  }, [models, providerAlias, provider.id]);
+
+  const chatModels = providerModels.filter((m) => !m.type);
+  const embeddingModels = providerModels.filter((m) => m.type === "embedding");
+  const imageModels = providerModels.filter((m) => m.type === "image");
+
+  const renderModelGroup = (title: string, icon: string, groupModels: EndpointModelSummary[]) => {
+    if (groupModels.length === 0) return null;
+    return (
+      <div className="mb-4">
+        <h4 className="text-xs font-semibold text-text-muted uppercase tracking-wider mb-2 flex items-center gap-1.5">
+          <span className="material-symbols-outlined text-sm">{icon}</span>
+          {title} ({groupModels.length})
+        </h4>
+        <div className="flex flex-col gap-1">
+          {groupModels.map((m) => {
+            const copyKey = `modal-${m.id}`;
+            return (
+              <div
+                key={m.id}
+                className="flex items-center gap-2 px-3 py-2 rounded-lg hover:bg-surface/60 group"
+              >
+                <code className="text-sm font-mono flex-1 truncate">{m.id}</code>
+                {m.custom && (
+                  <span className="text-[10px] px-1.5 py-0.5 rounded bg-primary/10 text-primary">
+                    {t("custom")}
+                  </span>
+                )}
+                <button
+                  onClick={() => copy(m.id, copyKey)}
+                  className="p-1 hover:bg-sidebar rounded text-text-muted hover:text-primary opacity-0 group-hover:opacity-100 transition-opacity"
+                  title={tc("copy")}
+                >
+                  <span className="material-symbols-outlined text-sm">
+                    {copied === copyKey ? "check" : "content_copy"}
+                  </span>
+                </button>
+              </div>
+            );
+          })}
+        </div>
+      </div>
+    );
+  };
+
+  return (
+    <Modal
+      isOpen
+      onClose={onClose}
+      title={t("providerModelsTitle", { provider: provider.provider.name })}
+    >
+      <div className="max-h-[60vh] overflow-y-auto">
+        {providerModels.length === 0 ? (
+          <p className="text-sm text-text-muted py-4 text-center">{t("noModelsForProvider")}</p>
+        ) : (
+          <>
+            {renderModelGroup(t("chat"), "chat", chatModels)}
+            {renderModelGroup(t("embedding"), "data_array", embeddingModels)}
+            {renderModelGroup(t("image"), "image", imageModels)}
+          </>
+        )}
+      </div>
+    </Modal>
+  );
+}
